+++ apps/core/src/index.ts
@@ -15,7 +15,6 @@ export * from './hooks/useGetObject';
 export * from './hooks/useGetDynamicFieldObject';
 export * from './hooks/useGetDynamicFields';
 export * from './hooks/useGetNormalizedMoveStruct';
-export * from './PostHogAnalyticsProvider';
 export * from './utils/formatPercentageDisplay';
 export * from './hooks/useGetSystemState';
 export * from './hooks/useGetTimeBeforeEpochNumber';
