+++ apps/explorer/package.json
@@ -44,7 +44,6 @@
         "clsx": "^1.2.1",
         "framer-motion": "^10.11.6",
         "plausible-tracker": "^0.3.8",
-        "posthog-js": "^1.53.1",
         "prism-react-renderer": "^1.3.5",
         "prism-themes": "^1.9.0",
         "prismjs": "^1.29.0",
